New Robot Python 3
Convertion with 2to3 and updates

+++ Scripts/Python/xRobot/link.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from Plasma import *
 
 #
@@ -7,7 +8,7 @@ def LinkPlayerTo(age, playerID = None, spawnPointNumber = None):
         playerID = PtGetLocalPlayer().getPlayerID()
     else:
         try:
-            playerID = long(playerID)
+            playerID = int(playerID)
         except:
             return "incorrect playerID"
             #pass
